Revert some now-unneeded changes to testing scripts.
